chore(trading): use custom genesis
