+++ apps/trading/e2e/tests/market_lifecycle/test_market_lifecycle.py
@@ -74,9 +74,13 @@ def test_market_lifecycle(proposed_market, vega: VegaServiceNull, page: Page):
         wallet=vega.wallet,
     )
 
+    vega.wait_fn(1)
+    vega.wait_for_total_catchup()
     # "wait" for market to be approved and enacted
     next_epoch(vega=vega)
+    next_epoch(vega=vega)
     # check that market is in pending state
+    page.pause()
     expect(trading_mode).to_have_text("Opening auction")
     expect(market_state).to_have_text("Pending")
