+++ apps/trading/e2e/tests/teams/test_teams.py
@@ -245,7 +245,7 @@ def test_team_page_games_table(team_page: Tuple[Page, str, str, VegaServiceNull]
     page.get_by_test_id("games-toggle").click()
     expect(page.get_by_test_id("games-toggle")).to_have_text("Results (10)")
     expect(page.get_by_test_id("rank-0")).to_have_text("1")
-    expect(page.get_by_test_id("epoch-0")).to_have_text("19")
+    expect(page.get_by_test_id("epoch-0")).to_have_text("18")
     expect(page.get_by_test_id("endtime-0")).to_be_visible()
     expect(page.get_by_test_id("type-0")).to_have_text("Price maker fees paid • tDAI ")
     expect(page.get_by_test_id("asset-0")).to_have_text("VEGA")
@@ -265,7 +265,7 @@ def test_team_page_members_table(team_page: Tuple[Page, str, str, VegaServiceNul
         "Owner"
     )
     expect(page.get_by_test_id("joinedAt-0")).to_be_visible()
-    expect(page.get_by_test_id("joinedAtEpoch-0")).to_have_text("9")
+    expect(page.get_by_test_id("joinedAtEpoch-0")).to_have_text("8")
 
 
 def test_team_page_headline(team_page: Tuple[Page, str, str, VegaServiceNull]):
@@ -326,7 +326,7 @@ def test_game_results_page(competitions_page: Tuple[Page, str, VegaServiceNull])
     page.goto(COMPETITIONS_URL)
     page.get_by_test_id("active-rewards-card").click()
     expect(page.get_by_test_id("dispatch-metric-info").first).to_have_text("Price maker fees paid • tDAI")
-    expect(page.get_by_test_id("epoch-0")).to_have_text("19")
+    expect(page.get_by_test_id("epoch-0")).to_have_text("18")
     expect(page.get_by_test_id("rank-0")).to_have_text("1")
     expect(page.get_by_test_id("teamName-0")).to_be_visible()
     expect(page.get_by_test_id("amount-0")).to_have_text("50.00")
