+++ apps/trading/e2e/vegahome/genesis.json
@@ -141,7 +141,7 @@
       "market.auction.maximumDuration": "168h",
       "market.auction.minimumDuration": "1s",
       "market.fee.factors.infrastructureFee": "0.0005",
-      "market.fee.factors.makerFee": "0.0002",
+      "market.fee.factors.makerFee": "0.1",
       "market.liquidity.bondPenaltyParameter": "0.1",
       "market.liquidity.earlyExitPenalty": "0.1",
       "market.liquidity.maximumLiquidityFeeFactorLevel": "0.03",
@@ -182,6 +182,8 @@
       "reward.staking.delegation.payoutDelay": "0h",
       "reward.staking.delegation.payoutFraction": "1.0",
       "rewards.marketCreationQuantumMultiple": "1",
+      "rewards.vesting.benefitTiers": "{\"tiers\": [{\"minimum_quantum_balance\": \"10000000\", \"reward_multiplier\": \"2\"}]}",
+      "rewards.activityStreak.benefitTiers": "{\"tiers\": [{\"minimum_activity_streak\": 2, \"reward_multiplier\": \"2.0\", \"vesting_multiplier\": \"1.1\"}]}",
       "snapshot.interval.length": "300",
       "spam.pow.difficulty": "15",
       "spam.pow.hashFunction": "sha3_24_rounds",
